requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.26.2
flask-cors==4.0.0
flask-mysql==1.6.0
flask-sqlalchemy==3.1.1
